add updateRange.js
